+++ artifactory/commands/utils/result_test.go
@@ -1,19 +1,23 @@
 package utils
 
 import (
+	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
+	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
 	"github.com/jfrog/jfrog-cli-core/v2/utils/tests"
 	clientutils "github.com/jfrog/jfrog-client-go/utils"
 	"github.com/stretchr/testify/assert"
 	"io"
 	"os"
 	"path"
+	"path/filepath"
+	"strconv"
 	"strings"
 	"testing"
 )
 
 // Check a case that targetRepository is not written in a deployableArtifacts file and needs to be read from the config file.
 func TestUnmarshalDeployableArtifacts(t *testing.T) {
-	err, cleanUpJfrogHome := tests.CreateTempJfrogConfig()
+	err, cleanUpJfrogHome := CreateTempJfrogConfig()
 	defer cleanUpJfrogHome()
 	// DeployableArtifact file is being change during runtime so a copy needs to be created.
 	tempDeployableArtifacts, err := createTempDeployableArtifactFile()
@@ -42,4 +46,51 @@ func createTempDeployableArtifactFile() (string, error) {
 
 func getTestsDataGradlePath() string {
 	return path.Join("..", "testdata", "gradle")
+}
+
+func CreateTempJfrogConfig() (err error, cleanUp func()) {
+	err, cleanUp = tests.SetJfrogHome()
+	if err != nil {
+		return err, nil
+	}
+	homedir := os.Getenv(coreutils.HomeDir)
+
+	configuration := `
+		{
+		  "artifactory": [
+			{
+			  "url": "http://localhost:8080/artifactory/",
+			  "user": "user",
+			  "password": "password",
+			  "serverId": "name",
+			  "isDefault": true
+			},
+			{
+			  "url": "http://localhost:8080/artifactory/",
+			  "user": "user",
+			  "password": "password",
+			  "serverId": "notDefault"
+			}
+		  ],
+		  "version": "2"
+		}
+	`
+	content, err := config.ConvertIfNeeded([]byte(configuration))
+	if err != nil {
+		return
+	}
+
+	versionString := ".v" + strconv.Itoa(coreutils.GetConfigVersion())
+	configFile, err := os.Create(filepath.Join(homedir, coreutils.JfrogConfigFile+versionString))
+
+	//configFile, err := os.Create(homedir)
+	if err != nil {
+		return
+	}
+	defer configFile.Close()
+	_, err = configFile.Write(content)
+	if err != nil {
+		return
+	}
+	return
 }
